fix error raised in MissionsManager.cpp:397 - Assert(pointer);
The error occurred because the LiveService service was not created.
Most likely this was due to the fact that 2 files had the same names and
CRT initialization took wrong `LiveService.cpp` file. The problem was
reproduced only in Debug build - I think that due to the optimization of Release build
the fully commented out file was immediately thrown away.